backend/requirements.txt
# API
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.6.4

# Database
sqlalchemy==2.0.29
aiosqlite==0.20.0

# Vector DB
chromadb==0.4.24

# Embeddings
sentence-transformers==2.6.1
torch==2.2.2

# Document loaders
pymupdf==1.24.1
beautifulsoup4==4.12.3
gitpython==3.1.43
markdown==3.6
html2text==2024.2.26

# Chunking
tiktoken==0.6.0

# LLM
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
tenacity==8.2.3
structlog==24.1.0

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
